pls work

+++ index.js
@@ -11,34 +11,98 @@ const prefix = process.env.prefix;
 let bugCooldowns = new Map();
 let announceCooldowns = new Map();
 let purgeCooldowns = new Map();
+let invite = 'https://discord.com/api/oauth2/authorize?client_id=743529355107500033&permissions=8&scope=bot';
 
 function formatNumber(x) {
     return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
 }
-function help(message){
+function help(message,args){
     let embed = new Discord.MessageEmbed()
         .setColor('#0099ff')
         .setTitle('Help')
-        .setAuthor('Mineware Bot')
-        .addFields(
-            {name: `How to use the command`, value: `Type ${prefix}help [section] to get the commands in that section. For example, ${prefix}help misc`, inline: false}
-            {name: 'help', value: 'Brings up this message', inline: false},
-            {name: 'hi', value: 'Says hello', inline: false},
-            {name: 'purge', value: 'Deletes some amount of messages. Requires the \`MANAGE_MESSAGES\` permission', inline: false},
-            {name: 'stats', value: 'Brings up stats like the server members, uptime, etc', inline: false},
-            {name: 'covid', value: 'Gets the Covid-19 stats for the US or a state when given a statecode', inline: false},
-            {name: 'bug', value: 'Submit a bug', inline: false},
-            {name: 'announce', value: 'Send an announcement! (requires \`MANAGE_CHANNELS\` permission)'}
-        )
+        .setAuthor(`${client.user.username}`)
         .setTimestamp()
         .setFooter(`Type ${prefix}bug to report any bugs!`, client.user.avatar_url);
+    let argsu = []; // uppercase args
+    for (let i = 0; i < args.length; i++){
+        argsu.push(args[i].toString().toUpperCase());
+    }
+    if (!argsu){
+        embed.addFields(
+            {name: `How to use the command`, value: `Type ${prefix}help [section] to get the commands in that section. For example, ${prefix}help misc`, inline: false},
+            {name: `Text commands`, value: `Text commands (no paramaters)`, inline = true},
+            {name: `Moderation`, value: `Moderation commands (needs special permissions)`, inline = true},
+            {name: `Bug Report`, value: `Report bugs`, inline: true}
+        );
+    }
+    switch (argsu[0]){
+        case 'MODERATION':
+            embed.addFields(
+                {name: `purge`, value: `Deletes messages`, inline: false},
+                {name: `announce`, value: `Make an announcement`, inline: false}
+            );
+            break;
+        case 'TEXT COMMANDS':
+            embed.addFields(
+                {name: `hi`, value: `Hello!`, inline: false},
+                {name: `stats`, value: `Provides some bot stats`, inline: false},
+                {name: `covid`, value: `Gets covid stats for a US state or US`, inline: false}
+            )
+            break;
+        case 'BUG REPORT':
+            embed.addFields(
+                {name: `bug`, value: `Report a bug!`, inline: false}
+            )
+            break;
+        case 'PURGE':
+            embed.addFields(
+                {name: `Arguments`, value: `Arguments: amt (int)`, inline:false},
+                {name: `Required Permissions`, value: `None`, inilne: false}
+            )
+            break;
+        case 'ANNOUNCE':
+            embed.addFields(
+                {name: `announce`, value: `Arguments: Channel, args (words of the message)`, inline: false},
+                {name: `Required Permissions`, value: `\`MANAGE_CHANNELS\``, inline: false}
+            )
+            break;
+        case 'HI':
+            embed.addFields(
+                {name: `hi`, value: `Arguments: None`, inline: false},
+                {name: `Required Permissions`, value: `None`, inline: false}
+            )
+            break;
+        case 'STATS':
+            embed.addFields(
+                {name: `stats`, value: `Arguments: None`, inline: false},
+                {name: `Required Permissions`, value: `None`, inline: false}
+            )
+            break;
+        case 'COVID':
+            embed.addFields(
+                {name: `covid`, value: `Arguments: statecode (optional)`, inline: false},
+                {name: `Required Permissions`, value: `None`, inline: false}
+            )
+            break;
+        case 'BUG':
+            embed.addFields(
+                {name: `bug`, value: `Arguments: args (words)`, inline: false},
+                {name: `Required Permissions`, value: `None`, inline: false}
+            )
+            break;
+    }
+        
     message.channel.send(embed);
 }
 function hello(message){
     message.channel.send('Hello!');
 }
 function purge(message,args){
-    if (!message.member.hasPermission('MANAGE_MESSAGES')) return;
+    if (!message.member.hasPermission('MANAGE_MESSAGES')){
+        message.channel.send(`You need the \`MANAGE_MESSAGE\` permission to use this command`);
+        return;
+    }
+    if (!args[0]) return;
     let time = Math.floor(Date.now() / 1000);
     let name = message.member.user.tag;
     if (purgeCooldowns[name] == null){
@@ -84,7 +148,7 @@ function stats(message){
     const embed = new Discord.MessageEmbed()
         .setColor('#0099ff')
         .setTitle('Stats')
-        .setAuthor('Mineware Bot')
+        .setAuthor(`${client.user.username}`)
         .addFields(
             {name: 'Members', value: memberCount, inline: false},
             {name: 'Ping', value: ping, inline: false}
@@ -96,9 +160,13 @@ function stats(message){
 }
 async function covid(message, args){ // sends a discord.MessageEmbed
     let url = `https://api.covidtracking.com/v1/us/current.json`;
-    if (args.length > 0) {
+    let things = []; // things to get in api
+    let isState = false;
+    if (args.length > 0 && args[0].length == 2){
         url = `https://api.covidtracking.com/v1/states/${args[0].toLowerCase()}/current.json`;
+        isState = true;
     }
+    if (isState) console.log('is state');
     try{
         let json = await fetch(url);
         let data = await json.json();
@@ -108,7 +176,7 @@ async function covid(message, args){ // sends a discord.MessageEmbed
         let covidEmbed = new Discord.MessageEmbed()
             .setColor('#0099ff')
             .setTitle('Covid-19 Stats')
-            .setAuthor('Mineware Bot')
+            .setAuthor(`${client.user.username}`)
             .addFields(
                 {name: 'Positive:', value: formatNumber(data.positive), inline: false},
                 {name: 'Negative:', value: formatNumber(data.negative), inline: false}
@@ -127,6 +195,7 @@ async function covid(message, args){ // sends a discord.MessageEmbed
 }
 
 function bug(message,args){
+    if (args.length == 0) return;
     let time = Math.floor(Date.now() / 1000);
     let name = message.member.user.tag;
     if (bugCooldowns[name] == null){
@@ -152,7 +221,10 @@ function bug(message,args){
 }
 
 function announce(message,args){
-    if (!message.member.hasPermission('MANAGE_CHANNELS')) return;
+    if (!message.member.hasPermission('MANAGE_CHANNELS')){
+        message.channel.send(`You need the \`MANGE_CHANNELS\` permission to use this command`);
+        return;
+    }
     let time = Math.floor(Date.now() / 1000);
     let name = message.member.user.tag;
     if (announceCooldowns[name] == null){
@@ -182,54 +254,49 @@ function announce(message,args){
     channel.send(msg);
 }
 
-function rules(message){
-    let aID = message.author.id;
-    var data = fs.readFileSync('rules.txt', 'utf8');
-    client.users.cache.get(aID.toString()).send(data);
-}
-
 client.on('ready', function(){
-    console.log('Logged in.')
+    console.log(`Logged in as ${client.user.name}`);
     client.user.setActivity(`${prefix}help | v${pkg.version}`);
 });
 
 client.on('message', function(message){
     if (!message.guild) return;
-    if (message.author.bot){
-        return;
-    }
-    if (message.member.roles.cache.some(role => role.name == 'Muted')){
-        message.delete();
-    }
+    if (message.author.bot) return;
+    if (message.member.roles.cache.some(role => role.name == 'Muted')) message.delete();
     if (!message.content.startsWith(prefix)) return;
+
     let command = message.content.split(' ')[0].slice(1);
     let args = message.content.split(' ').slice(1);
     switch (command.toUpperCase()){
         case 'HELP':
             help(message);
             break;
+
         case 'HI':
             hello(message);
             break;
+
         case 'PURGE':
             purge(message,args);
             break;
+            
         case 'STATS':
             stats(message);
             break;
+
         case 'COVID':
             covid(message,args);
             break;
+
         case 'BUG':
             bug(message,args);
+
         case 'ANNOUNCE':
             announce(message,args);
-            break;
-        case 'RULES':
-            rules(message);
             break;
     }
     console.log(`Author: ${message.author.toString()} Command: ${message.content}`);
 })
+
 console.log(pkg);
 client.login(token);
